+++ tests/Executor/ListsTest.php
@@ -217,7 +217,7 @@ class ListsTest extends TestCase
                 'data'   => ['nest' => null],
                 'errors' => [
                     [
-                        'debugMessage' => 'Cannot return null for non-nullable field DataType.test.',
+                        'debugMessage' => 'Cannot return null for non-nullable field "DataType.test".',
                         'locations'    => [['line' => 1, 'column' => 10]],
                     ],
                 ],
@@ -260,7 +260,7 @@ class ListsTest extends TestCase
                 'data'   => ['nest' => null],
                 'errors' => [
                     [
-                        'debugMessage' => 'Cannot return null for non-nullable field DataType.test.',
+                        'debugMessage' => 'Cannot return null for non-nullable field "DataType.test".',
                         'locations'    => [['line' => 1, 'column' => 10]],
                     ],
                 ],
@@ -368,7 +368,7 @@ class ListsTest extends TestCase
                 'data'   => ['nest' => ['test' => null]],
                 'errors' => [
                     [
-                        'debugMessage' => 'Cannot return null for non-nullable field DataType.test.',
+                        'debugMessage' => 'Cannot return null for non-nullable field "DataType.test".',
                         'locations'    => [['line' => 1, 'column' => 10]],
                     ],
                 ],
@@ -411,7 +411,7 @@ class ListsTest extends TestCase
                 'data'   => ['nest' => ['test' => null]],
                 'errors' => [
                     [
-                        'debugMessage' => 'Cannot return null for non-nullable field DataType.test.',
+                        'debugMessage' => 'Cannot return null for non-nullable field "DataType.test".',
                         'locations'    => [['line' => 1, 'column' => 10]],
                     ],
                 ],
@@ -527,7 +527,7 @@ class ListsTest extends TestCase
                 'data'   => ['nest' => null],
                 'errors' => [
                     [
-                        'debugMessage' => 'Cannot return null for non-nullable field DataType.test.',
+                        'debugMessage' => 'Cannot return null for non-nullable field "DataType.test".',
                         'locations'    => [['line' => 1, 'column' => 10]],
                     ],
                 ],
@@ -542,7 +542,7 @@ class ListsTest extends TestCase
                 'data'   => ['nest' => null],
                 'errors' => [
                     [
-                        'debugMessage' => 'Cannot return null for non-nullable field DataType.test.',
+                        'debugMessage' => 'Cannot return null for non-nullable field "DataType.test".',
                         'locations'    => [['line' => 1, 'column' => 10]],
                     ],
                 ],
@@ -579,7 +579,7 @@ class ListsTest extends TestCase
                 'data'   => ['nest' => null],
                 'errors' => [
                     [
-                        'debugMessage' => 'Cannot return null for non-nullable field DataType.test.',
+                        'debugMessage' => 'Cannot return null for non-nullable field "DataType.test".',
                         'locations'    => [['line' => 1, 'column' => 10]],
                     ],
                 ],
@@ -596,7 +596,7 @@ class ListsTest extends TestCase
                 'data'   => ['nest' => null],
                 'errors' => [
                     [
-                        'debugMessage' => 'Cannot return null for non-nullable field DataType.test.',
+                        'debugMessage' => 'Cannot return null for non-nullable field "DataType.test".',
                         'locations'    => [['line' => 1, 'column' => 10]],
                     ],
                 ],
@@ -660,7 +660,7 @@ class ListsTest extends TestCase
                 'data'   => ['nest' => null],
                 'errors' => [
                     [
-                        'debugMessage' => 'Cannot return null for non-nullable field DataType.test.',
+                        'debugMessage' => 'Cannot return null for non-nullable field "DataType.test".',
                         'locations'    => [['line' => 1, 'column' => 10]],
                     ],
                 ],