+++ tests/Executor/NonNullTest.php
@@ -547,7 +547,7 @@ class NonNullTest extends TestCase
             'data'   => ['syncNest' => null],
             'errors' => [
                 [
-                    'debugMessage' => 'Cannot return null for non-nullable field DataType.syncNonNull.',
+                    'debugMessage' => 'Cannot return null for non-nullable field "DataType.syncNonNull".',
                     'locations'    => [['line' => 4, 'column' => 11]],
                 ],
             ],
@@ -574,7 +574,7 @@ class NonNullTest extends TestCase
             'data'   => ['syncNest' => null],
             'errors' => [
                 [
-                    'debugMessage' => 'Cannot return null for non-nullable field DataType.promiseNonNull.',
+                    'debugMessage' => 'Cannot return null for non-nullable field "DataType.promiseNonNull".',
                     'locations'    => [['line' => 4, 'column' => 11]],
                 ],
             ],
@@ -602,7 +602,7 @@ class NonNullTest extends TestCase
             'data'   => ['promiseNest' => null],
             'errors' => [
                 [
-                    'debugMessage' => 'Cannot return null for non-nullable field DataType.syncNonNull.',
+                    'debugMessage' => 'Cannot return null for non-nullable field "DataType.syncNonNull".',
                     'locations'    => [['line' => 4, 'column' => 11]],
                 ],
             ],
@@ -630,7 +630,7 @@ class NonNullTest extends TestCase
             'data'   => ['promiseNest' => null],
             'errors' => [
                 [
-                    'debugMessage' => 'Cannot return null for non-nullable field DataType.promiseNonNull.',
+                    'debugMessage' => 'Cannot return null for non-nullable field "DataType.promiseNonNull".',
                     'locations'    => [['line' => 4, 'column' => 11]],
                 ],
             ],
@@ -772,10 +772,10 @@ class NonNullTest extends TestCase
                 'anotherPromiseNest' => null,
             ],
             'errors' => [
-                ['debugMessage' => 'Cannot return null for non-nullable field DataType.syncNonNull.', 'locations' => [['line' => 8, 'column' => 19]]],
-                ['debugMessage' => 'Cannot return null for non-nullable field DataType.syncNonNull.', 'locations' => [['line' => 19, 'column' => 19]]],
-                ['debugMessage' => 'Cannot return null for non-nullable field DataType.promiseNonNull.', 'locations' => [['line' => 30, 'column' => 19]]],
-                ['debugMessage' => 'Cannot return null for non-nullable field DataType.promiseNonNull.', 'locations' => [['line' => 41, 'column' => 19]]],
+                ['debugMessage' => 'Cannot return null for non-nullable field "DataType.syncNonNull".', 'locations' => [['line' => 8, 'column' => 19]]],
+                ['debugMessage' => 'Cannot return null for non-nullable field "DataType.syncNonNull".', 'locations' => [['line' => 19, 'column' => 19]]],
+                ['debugMessage' => 'Cannot return null for non-nullable field "DataType.promiseNonNull".', 'locations' => [['line' => 30, 'column' => 19]]],
+                ['debugMessage' => 'Cannot return null for non-nullable field "DataType.promiseNonNull".', 'locations' => [['line' => 41, 'column' => 19]]],
             ],
         ];
 
@@ -1021,7 +1021,7 @@ class NonNullTest extends TestCase
         $expected = [
             'errors' => [
                 [
-                    'debugMessage' => 'Cannot return null for non-nullable field DataType.syncNonNull.',
+                    'debugMessage' => 'Cannot return null for non-nullable field "DataType.syncNonNull".',
                     'locations'    => [['line' => 2, 'column' => 17]],
                 ],
             ],
@@ -1043,7 +1043,7 @@ class NonNullTest extends TestCase
         $expected = [
             'errors' => [
                 [
-                    'debugMessage' => 'Cannot return null for non-nullable field DataType.promiseNonNull.',
+                    'debugMessage' => 'Cannot return null for non-nullable field "DataType.promiseNonNull".',
                     'locations'    => [['line' => 2, 'column' => 17]],
                 ],
             ],