+++ CHANGELOG.md
@@ -10,6 +10,11 @@
 - Allow input types to be passed in from variables using \stdClass instead of associative arrays (#535)
 - Fixes parsing of string literals of the form \u0000 for code points in the range [128, 255] inclusive
 - Parse UTF-16 surrogate pairs within string literals 
+- Added quotes around `parentType.fieldName` in error message:
+```diff
+- Cannot return null for non-nullable field parentType.fieldName.
++ Cannot return null for non-nullable field "parentType.fieldName".
+```
 
 #### v0.13.5
 - Fix coroutine executor when using with promise (#486)